requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
requests>=2.31.0
pinecone>=5.0.0
python-dotenv>=1.0.0
alembic>=1.12.0

